requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.29
python-docx>=1.1
lxml>=5.2
python-multipart>=0.0.9
